backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.2
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
openai==1.35.13
pandas==2.2.2
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
redis==5.0.7
regex==2024.7.24
requests==2.32.3
rich==13.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.3
tiktoken==0.7.0
tqdm==4.66.4
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
youtube-transcript-api==0.6.2
